# practica1-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Tienda Online API",
        "description": "API REST para gestión de tienda construida con Vue.js + Laravel. Incluye autenticación Sanctum, CRUD de productos y categorías, pedidos y roles.",
        "contact": {
            "email": "admin@tienda.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan López"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@tienda.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Logout exitoso"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado con permisos",
                "operationId": "ae2f8c1ae1f9c5aee99c01f5d13da30b",
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "rol": {
                                            "type": "string",
                                            "enum": [
                                                "admin",
                                                "editor",
                                                "cliente"
                                            ]
                                        },
                                        "permisos": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos con filtros y paginación",
                "operationId": "2af0fdbcd7c972b8f9bd45123cd57cfb",
                "parameters": [
                    {
                        "name": "busqueda",
                        "in": "query",
                        "description": "Buscar por nombre o descripción",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "description": "Filtrar por categoría",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "precio_min",
                        "in": "query",
                        "description": "Precio mínimo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "precio_max",
                        "in": "query",
                        "description": "Precio máximo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de productos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear nuevo producto",
                "operationId": "434c8599d4fe9aaa5aee59a5be0cf3fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "stock"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Laptop Dell"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Laptop de alto rendimiento"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 1299.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "7f8c025cedd4418ed0240a21f7f9402a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del producto"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto existente",
                "operationId": "82ed71c67980b16764a81d295405c031",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "operationId": "075714328668dc259b9f6efb188ab787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}